streamlit_apps/requirements.txt
streamlit-ace==0.1.1
sqlparse==0.5.0
PyYAML==6.0.1
streamlit-elements==0.1.*
st-diff-viewer==1.0.6
